feat: add endpoint to create transaction

+++ src/main/kotlin/kastro/dev/repositories/ClientRepository.kt
@@ -4,13 +4,13 @@ import kastro.dev.exceptions.ResourceNotFoundException
 import kastro.dev.models.Client
 
 class ClientRepository {
-    fun updateBalance(clientId: Int, newBalance: Int) {
+    fun updateBalance(clientId: Int, newBalance: Long) {
         val connection = Database.connection
 
         val sql = "UPDATE clientes SET saldo = ? WHERE id = ?"
         val query = connection.prepareStatement(sql)
 
-        query.setInt(1, newBalance)
+        query.setLong(1, newBalance)
         query.setInt(2, clientId)
 
         query.executeUpdate()
@@ -30,8 +30,8 @@ class ClientRepository {
             return Client.restore(
                 id = result.getInt("id"),
                 name = result.getString("nome"),
-                limit = result.getInt("limite"),
-                balance = result.getInt("saldo")
+                limit = result.getLong("limite"),
+                balance = result.getLong("saldo")
             )
         }
