+++ src/main/kotlin/kastro/dev/usecases/outputs/CreateTransactionOutput.kt
@@ -1,6 +1,6 @@
 package kastro.dev.usecases.outputs
 
 data class CreateTransactionOutput(
-    val limit: Int,
-    val balance: Int,
+    val limit: Long,
+    val balance: Long,
 )
